NU-2431 switch native glide data grid delete event
