+++ designer/client/src/components/modals/TestingDataRecords/Table.tsx
@@ -270,6 +270,27 @@ export const Table: React.FC<TableProps> = ({
         }
     }, []);
 
+    const rowsFromSelection = useMemo(() => {
+        if (!selection) return [];
+        const rowsArr = selection.rows && typeof selection.rows.toArray === "function" ? selection.rows.toArray() : [];
+        if (rowsArr.length > 0) return rowsArr.slice().sort((a, b) => a - b);
+
+        const range = selection.current?.range;
+        if (range && typeof range.y === "number" && typeof range.height === "number") {
+            return Array.from({ length: range.height }, (_, i) => range.y + i);
+        }
+
+        return [];
+    }, [selection]);
+
+    const handleRemove = useCallback(() => {
+        if (!onCellDeleted || rowsFromSelection.length === 0) return;
+        onCellDeleted(rowsFromSelection);
+        clearSelection();
+    }, [clearSelection, onCellDeleted, rowsFromSelection]);
+
+    const selectedCount = rowsFromSelection.length;
+
     return (
         <Box
             sx={{
@@ -289,6 +310,11 @@ export const Table: React.FC<TableProps> = ({
                 }}
             >
                 <DataEditor
+                    onDelete={() => {
+                        handleRemove();
+
+                        return false;
+                    }}
                     ref={ref}
                     columns={tableColumns}
                     getCellContent={getCellContent}
@@ -324,9 +350,9 @@ export const Table: React.FC<TableProps> = ({
                 </CellMenu>
                 {selection.rows.length > 0 && (
                     <TableFooter
-                        selection={selection}
+                        selectedCount={selectedCount}
                         allRowsNumber={data.length}
-                        onDeleteRows={onCellDeleted}
+                        handleRemoveRows={handleRemove}
                         clearSelection={clearSelection}
                     />
                 )}
